Improved cache locality & fixed GameWorld.containsPoint
- Fixed GameWorld.containsPoint returned the opposite of what it should have.
- Reduced StaticEntity memory usage by a further 140,516 bytes on the Everard Island map (248,424 bytes compared with the 0.3.0 release). While this savings is essentially irrelevant to main memory, it should be more cache friendly.
- The world is now limited to 2048 columns and 2048 rows. The max positions for x and y are each 65,535. Entity size is limited to 127 world units, though that could easily be increased to 255 at the cost of a bit shift  per each entity width()/height() access.

+++ bubolo/src/bubolo/util/TileUtil.java
@@ -129,9 +129,9 @@ public abstract class TileUtil
 	 * @return an integer representing the correct adaptive tiling state for the specified
 	 *         tile, according to the adaptive tiling mechanism outlined on the project wiki.
 	 */
-	public static int getTilingState(Entity t, World w, Class<?>[] targetClasses)
+	public static byte getTilingState(Entity t, World w, Class<?>[] targetClasses)
 	{
-		int stateSum = 0;
+		byte stateSum = 0;
 
 		getEdgeMatches(tilingStateArray, t, w, targetClasses);
 		boolean[] edges = tilingStateArray;
